BN-822 cleanup
